Keyboard Complate
Keyboard Complate

+++ ColusClient/KeyMap.cs
@@ -2,8 +2,8 @@
     public class KeyMap{
         public string KeyName { get; }
         public int KeyCode { get; }
-        public int Weight { get; }
-        public KeyMap(string keyName, int keyCode, int weight)
+        public float Weight { get; }
+        public KeyMap(string keyName, int keyCode, float weight)
         {
             this.KeyName = keyName;
             this.KeyCode = keyCode;
